+++ nmnb-application/src/main/kotlin/nmnb/application/config/QueryDslConfig.kt
@@ -1,0 +1,18 @@
+package nmnb.application.config
+
+import com.querydsl.jpa.impl.JPAQueryFactory
+import jakarta.persistence.EntityManager
+import jakarta.persistence.PersistenceContext
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+class QueryDslConfig {
+    @PersistenceContext
+    lateinit var em: EntityManager
+
+    @Bean
+    fun jpaQueryFactory(): JPAQueryFactory {
+        return JPAQueryFactory(em)
+    }
+}
